Add test for snapshot
Need to ignore the test for now. Unfortunately the file is not
recognized as valid by tar-rs. An error is reported after trying to read
the first entry of the archive: "numeric field did not have utf-8 text".
This happens because the checksum of the first entry is invalid. It's
still to determine if it's a tar-rs bug, or the file is really invalid.